MSVC cannot convert guint64 to double
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code/trunk/gwyddion@10709 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ modules/file/pt3file.c
@@ -742,14 +742,14 @@ pt3file_scan_line_triggers(const PicoHarpFile *pt3file,
             n = (lineno == yres-1) ? lineno-1 : lineno;
             n = linetriggers[n+1].globaltime - linetriggers[n].globaltime;
             if (lineno % 2 == 0 || !pt3file->imaging.common.bidirectional) {
-                ltr->start = (ltr->globaltime + tstart*n*(1.0 - 2*a) + a*n);
-                ltr->stop = (ltr->globaltime + tstop*n*(1.0 - 2*a) + a*n);
+                ltr->start = (ltr->globaltime + (guint64)(tstart*n*(1.0 - 2*a) + a*n));
+                ltr->stop = (ltr->globaltime + (guint64)(tstop*n*(1.0 - 2*a) + a*n));
             }
             else {
-                ltr->start = (ltr->globaltime + (fstart - 1.0)*n*(1.0 - 2*a)
-                              + a*n);
-                ltr->stop = (ltr->globaltime + (fstop - 1.0)*n*(1.0 - 2*a)
-                             + a*n);
+                ltr->start = (ltr->globaltime + (guint64)((fstart - 1.0)*n*(1.0 - 2*a)
+                              + a*n));
+                ltr->stop = (ltr->globaltime + (guint64)((fstop - 1.0)*n*(1.0 - 2*a)
+                             + a*n));
             }
         }
     }
@@ -760,8 +760,8 @@ pt3file_scan_line_triggers(const PicoHarpFile *pt3file,
             LineTrigger *ltr = linetriggers + lineno;
             n = (lineno == yres-1) ? lineno-1 : lineno;
             n = linetriggers[n+1].globaltime - linetriggers[n].globaltime;
-            ltr->start = ltr->globaltime + a*n;
-            ltr->stop = ltr->globaltime + (1.0 - 2*a)*n;
+            ltr->start = ltr->globaltime + (guint64)(a*n);
+            ltr->stop = ltr->globaltime + (guint64)((1.0 - 2*a)*n);
         }
     }
